Crate clap_markdown

Autogenerate Markdown documentation for clap command-line tools

See Examples for examples of the content clap-markdown generates.

Structs

MarkdownOptions

Options to customize the structure of the output Markdown document.

Functions

help_markdown

Format the help information for command as Markdown.

help_markdown_command

Format the help information for command as Markdown.

help_markdown_command_custom

Format the help information for command as Markdown, with custom options.

help_markdown_custom

Format the help information for command as Markdown, with custom options.

print_help_markdown

Format the help information for command as Markdown and print it.